+++ js-lib/baby-fs/native-browser-fs.ts
@@ -8,6 +8,7 @@ import {
   FILE_ALREADY_EXISTS_ERROR,
   FILE_NOT_FOUND_ERROR,
   NATIVE_BROWSER_PERMISSION_ERROR,
+  NATIVE_BROWSER_USER_ABORTED_ERROR,
   NOT_A_DIRECTORY_ERROR,
   UPSTREAM_ERROR,
 } from './error-codes';
@@ -398,8 +399,14 @@ export async function pickADirectory() {
     return dirHandle;
   } catch (err) {
     if (err instanceof Error) {
-      console.error(err);
+      if (err instanceof DOMException && err.name === 'AbortError') {
+        throw new NativeBrowserFileSystemError(
+          'The user aborted.',
+          NATIVE_BROWSER_USER_ABORTED_ERROR,
+        );
+      }
       throw new Error(err.message);
+      console.error(err);
     }
     throw err;
   }
@@ -418,7 +425,7 @@ export async function requestNativeBrowserFSPermission(
 }
 
 export function supportsNativeBrowserFs() {
-  if ('showOpenFilePicker' in window) {
+  if (typeof window !== 'undefined' && 'showOpenFilePicker' in window) {
     return true;
   } else {
     return false;