+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Nov 14, 2021
+
+- **New workspace:** We have refined the new workspace UI.
+
 ### Nov 13, 2021
 
 - **UI Redesign**: Bangle.io now has a more polished and consistent UI.